Move IdStrategy to new path Update README
