+++ src/Manager/Bulk/AbstractDbalWriteExecutor.php
@@ -13,10 +13,10 @@ use Doctrine\DBAL\Platforms\PostgreSQLPlatform;
 use Elrise\Bundle\DbalBundle\Config\BundleConfigurationInterface;
 use Elrise\Bundle\DbalBundle\Config\DbalBundleConfig;
 use Elrise\Bundle\DbalBundle\DBAL\DbalParameterType;
+use Elrise\Bundle\DbalBundle\Enum\DbalConfigurableExecutorInterface;
+use Elrise\Bundle\DbalBundle\Enum\IdStrategy;
 use Elrise\Bundle\DbalBundle\Exception\UniqueConstraintViolationException;
 use Elrise\Bundle\DbalBundle\Exception\WriteDbalException;
-use Elrise\Bundle\DbalBundle\Manager\Contract\DbalConfigurableExecutorInterface;
-use Elrise\Bundle\DbalBundle\Manager\Contract\IdStrategy;
 use Elrise\Bundle\DbalBundle\Sql\Builder\SqlBuilderInterface;
 use Elrise\Bundle\DbalBundle\Utils\IdGenerator;
 use Exception;
