+++ src/Manager/Bulk/BulkInserter.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Elrise\Bundle\DbalBundle\Manager\Bulk;
 
-use Elrise\Bundle\DbalBundle\Manager\Contract\BulkInserterInterface;
+use Elrise\Bundle\DbalBundle\Enum\BulkInserterInterface;
 use Exception;
 use InvalidArgumentException;
